+++ rotkehlchen/exchanges/exchange.py
@@ -366,6 +366,36 @@ class ExchangeInterface(CacheableMixIn, LockableQueryMixIn):
         an actual exchange query.
         """
         log.debug(f'Querying deposits/withdrawals history for {self.name} exchange')
+        if only_cache is False:
+            ranges = DBQueryRanges(self.db)
+            location_string = f'{self.location!s}_asset_movements_{self.name}'
+            with self.db.conn.read_ctx() as cursor:
+                ranges_to_query = ranges.get_location_query_ranges(
+                    cursor=cursor,
+                    location_string=location_string,
+                    start_ts=start_ts,
+                    end_ts=end_ts,
+                )
+
+            for query_start_ts, query_end_ts in ranges_to_query:
+                log.debug(
+                    f'Querying online deposits/withdrawals for {self.name} between '
+                    f'{query_start_ts} and {query_end_ts}',
+                )
+                new_movements = self.query_online_deposits_withdrawals(
+                    start_ts=query_start_ts,
+                    end_ts=query_end_ts,
+                )
+                with self.db.user_write() as write_cursor:
+                    if len(new_movements) != 0:
+                        self.db.add_asset_movements(write_cursor, new_movements)
+                    ranges.update_used_query_range(
+                        write_cursor=write_cursor,
+                        location_string=location_string,
+                        queried_ranges=[(query_start_ts, query_end_ts)],
+                    )
+
+        # Read all asset movements from the DB
         filter_query = AssetMovementsFilterQuery.make(
             from_ts=start_ts,
             to_ts=end_ts,
@@ -375,39 +405,8 @@ class ExchangeInterface(CacheableMixIn, LockableQueryMixIn):
             asset_movements = self.db.get_asset_movements(
                 cursor=cursor,
                 filter_query=filter_query,
-                has_premium=True,  # is okay since the returned trades don't make it to the user
+                has_premium=True,  # is okay since the returned events don't make it to the user
             )
-            if only_cache:
-                return asset_movements
-
-            ranges = DBQueryRanges(self.db)
-            location_string = f'{self.location!s}_asset_movements_{self.name}'
-            ranges_to_query = ranges.get_location_query_ranges(
-                cursor=cursor,
-                location_string=location_string,
-                start_ts=start_ts,
-                end_ts=end_ts,
-            )
-
-        for query_start_ts, query_end_ts in ranges_to_query:
-            log.debug(
-                f'Querying online deposits/withdrawals for {self.name} between '
-                f'{query_start_ts} and {query_end_ts}',
-            )
-            new_movements = self.query_online_deposits_withdrawals(
-                start_ts=query_start_ts,
-                end_ts=query_end_ts,
-            )
-
-            with self.db.user_write() as write_cursor:
-                if len(new_movements) != 0:
-                    self.db.add_asset_movements(write_cursor, new_movements)
-                ranges.update_used_query_range(
-                    write_cursor=write_cursor,
-                    location_string=location_string,
-                    queried_ranges=[(query_start_ts, query_end_ts)],
-                )
-            asset_movements.extend(new_movements)
 
         return asset_movements
 
@@ -424,41 +423,38 @@ class ExchangeInterface(CacheableMixIn, LockableQueryMixIn):
         an actual exchange query.
         """
         db = DBHistoryEvents(self.db)
+        if only_cache is False:
+            ranges = DBQueryRanges(self.db)
+            location_string = f'{self.location!s}_history_events_{self.name}'
+            with self.db.conn.read_ctx() as cursor:
+                ranges_to_query = ranges.get_location_query_ranges(
+                    cursor=cursor,
+                    location_string=location_string,
+                    start_ts=start_ts,
+                    end_ts=end_ts,
+                )
+
+            for query_start_ts, query_end_ts in ranges_to_query:
+                new_events = self.query_online_income_loss_expense(
+                    start_ts=query_start_ts,
+                    end_ts=query_end_ts,
+                )
+                with self.db.user_write() as write_cursor:
+                    if len(new_events) != 0:
+                        db.add_history_events(write_cursor, new_events)
+                    ranges.update_used_query_range(
+                        write_cursor=write_cursor,
+                        location_string=location_string,
+                        queried_ranges=[(query_start_ts, query_end_ts)],
+                    )
+
         filter_query = HistoryEventFilterQuery.make(
             from_ts=start_ts,
             to_ts=end_ts,
             location=self.location,
         )
         with self.db.conn.read_ctx() as cursor:
-            # has_premium True is fine here since the result of this is not user facing atm
             events = db.get_history_events(cursor, filter_query=filter_query, has_premium=True)
-            if only_cache:
-                return events  # type: ignore[return-value]  # HistoryBaseEntry vs HistoryEvent
-
-            ranges = DBQueryRanges(self.db)
-            location_string = f'{self.location!s}_history_events_{self.name}'
-            ranges_to_query = ranges.get_location_query_ranges(
-                cursor=cursor,
-                location_string=location_string,
-                start_ts=start_ts,
-                end_ts=end_ts,
-            )
-
-        for query_start_ts, query_end_ts in ranges_to_query:
-            new_events = self.query_online_income_loss_expense(
-                start_ts=query_start_ts,
-                end_ts=query_end_ts,
-            )
-            with self.db.user_write() as write_cursor:
-                if len(new_events) != 0:
-                    db.add_history_events(write_cursor, new_events)
-                ranges.update_used_query_range(
-                    write_cursor=write_cursor,
-                    location_string=location_string,
-                    queried_ranges=[(query_start_ts, query_end_ts)],
-                )
-            events.extend(new_events)
-
         return events  # type: ignore[return-value]  # HistoryBaseEntry vs HistoryEvent
 
     def query_history_with_callbacks(
@@ -497,7 +493,7 @@ class ExchangeInterface(CacheableMixIn, LockableQueryMixIn):
                 only_cache=False,
             )
             if new_step_data is not None:
-                new_step_callback(f'Querying {exchange_name} ledger actions history')
+                new_step_callback(f'Querying {exchange_name} events history')
             self.query_income_loss_expense(
                 start_ts=start_ts,
                 end_ts=end_ts,